dev-requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.0.0
apyori==1.1.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
blinker==1.6.3
bpyutils==0.5.8
bs4==0.0.1
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.3
Django==4.2.6
EasyProcess==1.1
emoji==2.8.0
entrypoint2==1.1
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
Faker==19.8.0
filelock==3.12.4
fitz==0.0.1.dev2
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.43.1
frontend==0.0.3
frozenlist==1.4.0
future==0.18.3
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprofiler-official==1.0.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
keras==2.14.0
keyboard==0.13.5
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mlxtend==0.23.0
MouseInfo==0.1.3
mpl-finance==0.10.1
mplfinance==0.12.10b0
mss==9.0.1
multidict==6.0.4
mygene==3.2.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parsel==1.8.1
pathlib==1.0.1
patsy==0.5.3
Pillow==10.0.1
pip-review==1.3.0
pip-tools==7.3.0
pipupgrade==1.12.0
platformdirs==3.11.0
pooch==1.7.0
Protego==0.3.0
protobuf==4.24.4
prov==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyDispatcher==2.0.7
pydot==1.4.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pypng==0.20220715.0
pyproject_hooks==1.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
qrcode==7.4.2
queuelib==1.6.2
rdflib==7.0.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
Scrapy==2.11.0
seaborn==0.13.0
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
starlette==0.31.1
statsmodels==0.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tldextract==3.6.0
tomli==2.0.1
tqdm==4.66.1
traits==6.3.2
tushare==1.2.89
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
w3lib==2.1.2
websocket-client==0.57.0
Werkzeug==3.0.0
win32-setctime==1.1.0
wrapt==1.14.1
xgboost==2.0.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
